+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog  
 All notable changes to this project will be documented in this file.
 
+## [0.20.2]
+
+### Fixed
+- Mark customer card payment as always online payment
+
 ## [0.20.1]
 
 ### Added